+++ backend/models/serv_models.js
@@ -1,11 +1,11 @@
-const mongoose=require(`mongoose`)
+// const mongoose=require(`mongoose`)
 
-const {Schema}=mongoose
+// const {Schema}=mongoose
 
 
-const serv_schema= new Schema({
-    name:{type:String,required:true},
+// const serv_schema= new Schema({
+//     name:{type:String,required:true},
     
-    description:{type:string},
+//     description:{type:string},
     
-})
+// })
